temp: moving req. status to custom hook
